Move SelectAction to BranchView.hpp
