--- infra/BranchViewAction.cpp
@@ -1,5 +1,0 @@
-//
-// Created by eugene on 25/09/2020.
-//
-
-#include "BranchViewAction.hpp"
